Added Passion Project via upload
The passion project selected was a Text-Adventure Game created with Python. As a avid reader, I decided to create a game based around the first book of the Maze Runner saga. The game consists of a 3-stage play, with multiple questions and endings, some in failure and some in victory. The concepts used were as follows:
Strings and String Manipulation
User-Defined Functions and Objects (variables)
User Input and Input Handling
Conditional and Nested Conditional Statements (if/elif/else)
Loops (for/while)
Controlling for Errors (try/except)