Added Regression Modeling Project via upload
For this regression project, the task at hand was developing a machine learning model to predict the number of bike rentals on a given day, as well as to provide insights into the factors that contribute to bike rental demand. We had a cohort competition in Kaggle, submitting various models and trying to get the highest score for accuracy. The permitted model types are as follows:
OLS Linear Regression	
Lasso Regression
Ridge Regression
Elastic Net Regression
K-Nearest Neighbors
Decision Tree Regressor